requirements.txt
Flask==2.3.3
tensorflow==2.18.0
scikit-learn==1.3.0
opencv-python==4.8.1.78
scikit-image==0.21.0
numpy==1.24.3
Pillow==10.0.1
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
Werkzeug==2.3.7
